+++ source/tools/DumpBin/PeHeader.h
@@ -2,8 +2,8 @@
 
 #include <Windows.h>
 
-DWORD DosHeader(PBYTE Data, DWORD Size);
-DWORD FileHeader(PBYTE Data, DWORD Size);
-DWORD OptionlHeader(PBYTE Data, DWORD Size);
-DWORD DataDirectory(PBYTE Data, DWORD Size);
-DWORD SectionHeader(PBYTE Data, DWORD Size);
+DWORD DosHeader(PBYTE Data, LONGLONG Size);
+DWORD FileHeader(PBYTE Data, LONGLONG Size);
+DWORD OptionlHeader(PBYTE Data, LONGLONG Size);
+DWORD DataDirectory(PBYTE Data, LONGLONG Size);
+DWORD SectionHeader(PBYTE Data, LONGLONG Size);
